move getSuggestionsFromQuery.ts
